pyproject.toml
[tool.poetry]
name = "repl-nix-inbox-manager-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_inbox_manager_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
langchain = "^0.1.8"
openai = "^1.12.0"
pydantic = "^2.6.1"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
